# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Goth is a robust authentication API that supports both OAuth providers and traditional credentials-based authentication. It is designed to provide secure and flexible user authentication for various applications.",
        "title": "Goth",
        "contact": {
            "name": "Ankur Kumar",
            "url": "https://github.com/sirjager"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/auth/user": {
            "get": {
                "description": "Get Authenticated User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Fetch",
                "responses": {
                    "200": {
                        "description": "UserResponse",
                        "schema": {
                            "$ref": "#/definitions/UserResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Initiate Authenticated User Deletion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code if already have",
                        "name": "code",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "Partially Update User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "description": "Update User Params",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateUserParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UserResponse",
                        "schema": {
                            "$ref": "#/definitions/UserResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/verify": {
            "get": {
                "description": "Email Verification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify Email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email to verify",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email verification code if already have any",
                        "name": "code",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "UpdateUserParams": {
            "type": "object",
            "properties": {
                "currentPassword": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                },
                "pictureURL": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "blocked": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "pictureURL": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
        "UserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/User"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}